+++ python/love/commander/app.py
@@ -20,13 +20,15 @@
 
 """Main application, instantiates the aiohtttp app."""
 from aiohttp import web
+
 from .commands import create_app as create_cmd_app
-from .heartbeats import create_app as create_heartbeat_app
-from .salinfo import create_app as create_salinfo_app
-from .lovecsc import create_app as create_lovecsc_app
 from .efd import create_app as create_efd_app
-from .tcs import create_app as create_tcs_app
+from .heartbeats import create_app as create_heartbeat_app
 from .lfa import create_app as create_lfa_app
+from .lovecsc import create_app as create_lovecsc_app
+from .reports import create_app as create_reports_app
+from .salinfo import create_app as create_salinfo_app
+from .tcs import create_app as create_tcs_app
 
 
 def create_app(*args, **kwargs):
@@ -45,6 +47,7 @@ def create_app(*args, **kwargs):
     app.add_subapp("/efd/", create_efd_app())
     app.add_subapp("/tcs/", create_tcs_app())
     app.add_subapp("/lfa/", create_lfa_app())
+    app.add_subapp("/reports/", create_reports_app())
 
     app.add_subapp(
         "/salinfo/",