+++ CHANGELOG.rst
@@ -2,6 +2,18 @@
 Version History
 ===============
 
+v6.1.0
+------
+
+* Add M1M3 Bump Tests reports `<https://github.com/lsst-ts/LOVE-commander/pull/66>`_
+
+v6.0.6
+------
+
+* Hotfix for Jenkinsfile `<https://github.com/lsst-ts/LOVE-commander/pull/65>`_
+* Move docs creation to CI `<https://github.com/lsst-ts/LOVE-commander/pull/63>`_
+* Add ts_pre_commit_conf `<https://github.com/lsst-ts/LOVE-commander/pull/64>`_
+
 v6.0.5
 ------